// apps/desktop/src/features/main/panel/types.ts
import { GraphNodeType, RelationType } from '@tgim/types/graph';

export type GraphAdjacencyEntry = {
  nodeId: string;
  relationKind: RelationType;
  kindRuleId: string;
};

export interface GraphContext {
  adjacency: Record<string, GraphAdjacencyEntry[]>;
  nodeTypes: Record<string, GraphNodeType>;
  connectionKinds: RelationType[];
  availableLevels: number[];
  kindRuleIds: string[];
  nodeLabels: Record<string, string>;
}